rename file and fix outdated comments
